Created user middleware to check user against production db.
Removed the post route from the user route group and assign the req.user with information from the database. Also cleaned up variable names when pulling IDs from params.
