+++ src/app.js
@@ -3,12 +3,13 @@ const { userRouter, matcherRouter, partyRouter } = require('./routes')
 const express = require('express')
 const app = express()
 const { auth } = require('express-openid-connect')
+const { userAuth } = require('./middleware')
 
 const { AUTH0_SECRET, AUTH0_AUDIENCE, AUTH0_CLIENT_ID, AUTH0_BASE_URL } =
   process.env
 
 const config = {
-  authRequired: false,
+  authRequired: true,
   auth0Logout: true,
   secret: AUTH0_SECRET,
   baseURL: AUTH0_AUDIENCE,
@@ -20,6 +21,8 @@ app.use(auth(config))
 
 app.use(express.json())
 app.use(express.urlencoded({ extended: true }))
+
+app.use(userAuth)
 
 // User routes (mainly for manual testing)
 app.use('/users', userRouter)
